+++ StartPP/PrintHelper.cpp
@@ -81,7 +81,8 @@ void CPrintHelper::DrawPageBorder
 }
 
 
-void CPrintHelper::DrawFrame(CDC* pDC, CRect& rcDib, CString strFileName, double fAspX, double fAspY, const SFloatRect& margins)
+void CPrintHelper::DrawFrame(CDC *pDC, CRect &rcDib, CString strFileName, float fAspX, float fAspY,
+							 const SFloatRect &margins)
 {
 	wxSize sz = rcDib.GetSize();
